Improve documentation for #[test] attribute
The first version of the crate was very sparsely documented as a release
was urgently needed. With this change we start adding better
documentation for the #[test] attribute to make usage more obvious. To
have some basic sanity checking of the documentation, also add a CI job
that generates it.