core: Implement Display for RustyForkId directly
In order to make use of a RustyForkId for its intended purpose, it has
to be converted to a string. So far this conversion was happening
detached from the type definition, but conceptually it makes the most
sense to co-locate the two. Do just that. Then extend the tests to make
sure that the string representation also has the properties we desire.