core: Remove TestExitStatus trait
The TestExitStatus trait, while conceptually useful, on closer
inspection seems inadequate for the task at hand. Specifically, it is
implemented only for a sub-set of return types that tests may actually
have. All supported types seem to be covered by the
std::process::Termination trait. Thus, with this change we remove
TestExitStatus in favor of using Termination. In so doing we also remove
an implementation for the Option type, which isn't actually a supported
return type by test functions.