core: Remove tempfile dependency
So far we have used a temporary file for conveying the child's output to
the parent. However, that's actually unnecessary, because the same thing
can be achieved with only built-in means and without a 3rd party
dependency. To make that happen, we piggy back on the piping
functionality that comes with Rust's std::process::Command type, which
now also allows us to differentiate between stdout and stderr, instead
of clubbing both together.