core: Simplify signature of fork()
The fork() function is notoriously complex in that it accepts a bunch of
arguments that ultimately should be implementation/test details but
currently have to be provided by clients.
With this change we simplify it, by removing unnecessary arguments from
the public signature, while switching tests over to using the new
fork_int(), which still allows for the same level of configurability. In
the process, we also adjust the function running in the child (i.e., the
test function body) to stop supporting any kind of captured variable
mutation. We cannot support transparently modifying data across process
boundaries the way that the signature previously conveyed.