Convert some "fork_test" tests into end-to-end ones
The unit tests accompanying the fork_test module and testing the
fork_test macro really could and probably should be tests using our
version of the #[test] macro instead. Convert them. With that we start
to have some tests for what is our new public API (the #[test]
attribute), while removing a dependency on the fork_test macro, which
hopefully we can ditch soon.