core: Remove another user of fork_test macro
The command line logic uses the fork_test macro for part of its testing.
Switch this test over to using the fork() function instead, now that its
signature is much less unwieldy. In so doing we removed the last
crate-internal user of the fork_test macro, paving the way for its
eventual removal.