macros: Add support for #[bench] attribute
Add support for the #[bench] attribute, as a way to run benchmarks in a
separate process. The attribute basically behaves similar to the
existing #[test], but requires a different signature of the benchmark
function. Because benchmarks work with a mutable reference to a Bencher
object, we need to make sure to transfer this object to the child
process and then back. Unfortunately, given that the type is opaque from
a user's perspective, just treating it as a byte buffer is potentially
unsound. To make this somewhat obvious, introduce the 'unsound' feature.
Furthermore, because the entire concept of libtest based benchmarks is
nightly only, we put everything behind the 'unstable' as well.