core: Introduce fork_in_out()
Aside from stdout and stderr "redirection", it basically currently not
possible to exchange data with the test. However, doing so may be
required in certain circumstances.
To address this shortcoming, this change introduces the fork_in_out()
function, which support the passing of a byte buffer to the child and
back. Internally, the data will be transferred over a TCP stream
established for this purpose.